Addition of the work to allow Features to be added to the system. This used the basic MVC scaffolding but is not yet styled

+++ TeamCityMonitor/Models/Feature.cs
@@ -2,7 +2,7 @@
 {
     public class Feature
     {
-        public int FeatureId { get; set; }
+        public long FeatureId { get; set; }
         public string FeatureName { get; set; }
         public bool Enabled { get; set; }
     }
